% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildConnecteur.R
\name{createPostgreSQLConnector}
\alias{createPostgreSQLConnector}
\title{Création d'un connecteur postgreSQL pour l'application shiny}
\usage{
createPostgreSQLConnector(
  host,
  dbname,
  port = 5432,
  filename = "./PostgreSQLConnector.R"
)
}
\arguments{
\item{host}{adresse du serveur postgreSQL}

\item{dbname}{nom de la base de donnéespostgreSQL}

\item{port}{port du serveur postgreSQL}

\item{filename}{nom du fichier R de connexion}
}
\description{
Cette fonction permet de construire un fichier .R qui pourra être utilisé en source de l'application shiny.
}
