% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher.R
\name{ExplorerDonnees}
\alias{ExplorerDonnees}
\title{Lancement d'une appli Shiny en fond}
\usage{
ExplorerDonnees(connectorFile = NULL)
}
\arguments{
\item{connectorFile}{Nom et chemin du fichier de connexion. Le fichier de connexion doit permettre la construction
d'une liste de connecteur, nommé connectors.}
}
\description{
Lancement d'une appli Shiny en fond
}
