% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher.R
\name{launch_shiny_backgroundJob}
\alias{launch_shiny_backgroundJob}
\title{Lancement d'une appli Shiny en background}
\usage{
launch_shiny_backgroundJob(
  shinyPath,
  host = "127.0.0.1",
  port = NULL,
  name = NULL,
  opt_cmd = NULL,
  max_attempts = 30
)
}
\arguments{
\item{shinyPath}{le chemin de l'application shiny}

\item{host}{l'adresse de connexion à l'application}

\item{port}{le port d'ecoute}

\item{name}{nom du job}

\item{opt_cmd}{commande optionnelle à lancer avant le run de l'appl}
}
\value{
rien
}
\description{
Lancement d'une appli Shiny en background
}
