% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher.R
\name{launch_shiny_backgroundJob}
\alias{launch_shiny_backgroundJob}
\title{Lancement d'une appli Shiny en fond}
\usage{
launch_shiny_backgroundJob(
  shinyPath,
  port = NULL,
  sleep = 5,
  name = NULL,
  opt_cmd = NULL
)
}
\arguments{
\item{shinyPath}{le chemin de l'application shiny}

\item{port}{le port d'ecoute}

\item{sleep}{temp sleep avant lancement}

\item{name}{nom du job}

\item{opt_cmd}{commande optionnelle à lancer avant le run de l'appl}
}
\value{
rien
}
\description{
Lancement d'une appli Shiny en fond
}
